squash: refactor biging typedarray testing
It needs to be in its own file so we can avoid
loading it completely on earlier Node.js versions
