+++ test/index.js
@@ -31,6 +31,7 @@ let testModules = [
   'object/set_property',
   'promise',
   'typedarray',
+  'typedarray-bigint',
   'objectwrap',
   'objectreference',
   'version_management'
@@ -45,6 +46,7 @@ if((process.env.npm_config_NAPI_VERSION !== undefined) &&
   // this should be guarded on the napi version
   // in which bigint was added.
   testModules.splice(testModules.indexOf('bigint'), 1);
+  testModules.splice(testModules.indexOf('typedarray-bigint'), 1);
 }
 
 if (typeof global.gc === 'function') {
